fixes type instatiation is excessively deep error in sidebarSlice
