+++ src/sidebar/sidebarSlice.test.ts
@@ -15,10 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import sidebarSlice, {
-  addFormPanel,
-  fixActiveTabOnRemove,
-} from "@/sidebar/sidebarSlice";
+import sidebarSlice, { fixActiveTabOnRemove } from "@/sidebar/sidebarSlice";
 import { eventKeyForEntry } from "@/sidebar/eventKeyUtils";
 import {
   cancelForm,
@@ -33,6 +30,8 @@ import { uuidv4, validateRegistryId } from "@/types/helpers";
 import { MOD_LAUNCHER } from "@/sidebar/modLauncher/constants";
 import { type Draft } from "immer";
 import { waitFor } from "@testing-library/react";
+import { addFormPanel } from "@/sidebar/thunks";
+import { configureStore } from "@reduxjs/toolkit";
 
 jest.mock("@/sidebar/messenger/api");
 jest.mock("@/contentScript/messenger/strict/api");
@@ -216,13 +215,13 @@ describe("sidebarSlice.removeTemporaryPanel", () => {
 
 describe("sidebar thunk addFormPanel", () => {
   it("adds form to empty state", async () => {
-    const state = sidebarSlice.getInitialState();
+    const store = configureStore({
+      reducer: sidebarSlice.reducer,
+    });
 
     const extensionId = autoUUIDSequence();
 
-    const newState = sidebarSlice.reducer(
-      state,
-      // @ts-expect-error -- typings for thunk actions are not in a good place right now
+    await store.dispatch(
       addFormPanel({
         form: {
           type: "form",
@@ -242,10 +241,8 @@ describe("sidebar thunk addFormPanel", () => {
       }),
     );
 
-    await tick();
-
     await waitFor(() => {
-      expect(newState.forms).toHaveLength(1);
+      expect(store.getState().forms).toHaveLength(1);
     });
     expect(cancelFormMock).toHaveBeenCalledExactlyOnceWith({
       frameId: 0,
